1. BufferedImageCollector, ByteListCollector, Collector, FileCollector, StreamableImageCollector moved to pl.uj.edu.JImageStream.api.collectors 2. Wildcart imports removed 3. Runner refactored using Collector interface
