+++ src/main/java/pl/uj/edu/JImageStream/api/collectors/BufferedImageCollector.java
@@ -1,4 +1,4 @@
-package pl.uj.edu.JImageStream.api;
+package pl.uj.edu.JImageStream.api.collectors;
 
 import java.awt.image.BufferedImage;
