+++ src/main/java/pl/uj/edu/JImageStream/api/collectors/ByteListCollector.java
@@ -1,8 +1,7 @@
-package pl.uj.edu.JImageStream.api;
+package pl.uj.edu.JImageStream.api.collectors;
 
 import java.awt.image.BufferedImage;
 import java.awt.image.DataBufferByte;
-import java.awt.image.WritableRaster;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -10,9 +9,7 @@ import java.util.List;
 public class ByteListCollector implements Collector<List<Byte>> {
     @Override
     public List<Byte> collect(BufferedImage bufferedImage) {
-        WritableRaster writableRaster = bufferedImage.getRaster();
-        DataBufferByte dataBufferByte = (DataBufferByte) writableRaster.getDataBuffer();
-        byte[] data = dataBufferByte.getData();
+        byte[] data = ((DataBufferByte) bufferedImage.getRaster().getDataBuffer()).getData();
         List<Byte> list = new ArrayList<>();
         for (byte d : data) {
             list.add(d);
